Move agentWithDefaults back to obfuscate_test.go
